+++ packages/dartitect_mcp/lib/src/server.dart
@@ -28,7 +28,7 @@ base class DartitectMcpServer extends MCPServer
        super.fromStreamChannel(
          implementation: Implementation(
            name: 'dartitect_mcp',
-           version: '1.0.0-rc.4',
+           version: '1.0.0-rc.5',
          ),
          instructions: _instructions,
        ) {